+++ src/app.ts
@@ -7,7 +7,7 @@ import * as config from "./config";
 import * as type from './common/types';
 import {logger, logErr} from "./utils/logger";
 import * as mongo from "./mongo/mongo";
-import * as filters from "./utils/date-filter";
+import * as filters from "./utils/nunjuck-custom-filters";
 
 
 // Convert the zlib.unzip function to return a promise using promisify
